PER-960 thread safe stubs

+++ invesdwin-context-integration-parent/invesdwin-context-integration-ftp-server/src/main/java/de/invesdwin/context/integration/ftp/server/test/internal/FtpServerTestStub.java
@@ -2,7 +2,8 @@ package de.invesdwin.context.integration.ftp.server.test.internal;
 
 import java.util.List;
 
-import javax.annotation.concurrent.NotThreadSafe;
+import javax.annotation.concurrent.GuardedBy;
+import javax.annotation.concurrent.ThreadSafe;
 
 import org.apache.ftpserver.FtpServer;
 import org.springframework.beans.factory.NoSuchBeanDefinitionException;
@@ -23,10 +24,11 @@ import de.invesdwin.util.shutdown.ShutdownHookManager;
 import jakarta.inject.Named;
 
 @Named
-@NotThreadSafe
+@ThreadSafe
 public class FtpServerTestStub extends StubSupport {
 
-    private static volatile FtpServer lastServer;
+    @GuardedBy("this.class")
+    private static FtpServer lastServer;
 
     static {
         ShutdownHookManager.register(new IShutdownHook() {
@@ -40,8 +42,6 @@ public class FtpServerTestStub extends StubSupport {
 
     @Override
     public void setUpContextLocations(final ATest test, final List<PositionedResource> locations) throws Exception {
-        //if for some reason the tearDownOnce was not executed on the last test (maybe maven killed it?), then try to stop here aswell
-        maybeStopLastServer();
         final FtpServerTest annotation = Reflections.getAnnotation(test, FtpServerTest.class);
         if (annotation != null) {
             if (annotation.value()) {
@@ -54,6 +54,11 @@ public class FtpServerTestStub extends StubSupport {
 
     @Override
     public void setUpContext(final ATest test, final TestContext ctx) throws Exception {
+        if (ctx.isPreMergedContext()) {
+            return;
+        }
+        //if for some reason the tearDownOnce was not executed on the last test (maybe maven killed it?), then try to stop here aswell
+        maybeStopLastServer();
         //clean up for next test
         Files.deleteQuietly(FtpServerProperties.WORKING_DIRECTORY);
         Files.forceMkdir(FtpServerProperties.WORKING_DIRECTORY);
@@ -61,10 +66,14 @@ public class FtpServerTestStub extends StubSupport {
 
     @Override
     public void setUpOnce(final ATest test, final TestContext ctx) throws Exception {
-        try {
-            FtpServerTestStub.lastServer = MergedContext.getInstance().getBean(FtpServer.class);
-        } catch (final NoSuchBeanDefinitionException e) { //SUPPRESS CHECKSTYLE empty block
-            //ignore
+        synchronized (FtpServerTestStub.class) {
+            if (FtpServerTestStub.lastServer == null) {
+                try {
+                    FtpServerTestStub.lastServer = MergedContext.getInstance().getBean(FtpServer.class);
+                } catch (final NoSuchBeanDefinitionException e) { //SUPPRESS CHECKSTYLE empty block
+                    //ignore
+                }
+            }
         }
     }
 
@@ -76,7 +85,7 @@ public class FtpServerTestStub extends StubSupport {
         maybeStopLastServer();
     }
 
-    private static void maybeStopLastServer() throws Exception {
+    private static synchronized void maybeStopLastServer() throws Exception {
         if (lastServer != null) {
             lastServer.stop();
             lastServer = null;
